requirements.txt
streamlit==1.32.2
opencv-python==4.9.0.80
scikit-learn==1.4.1.post1
tensorflow==2.15.0
face_recognition==1.3.0
dlib
numpy==1.24.4
pickle5==0.0.11
matplotlib==3.8.2
Pillow==10.2.0